requirements.txt
opencv_python==4.5.1.48
PyAutoGUI==0.9.52
numpy==1.19.3
paddlex==1.3.7
Pillow==8.1.2
